# refactor(SnowflakeUtil): clean up utils and improve perf