--- a/src/util/Util.js
+++ b/src/util/Util.js
@@ -523,57 +523,6 @@ class Util extends null {
   }
 
   /**
-   * Transforms a snowflake from a decimal string to a bit string.
-   * @param {Snowflake} num Snowflake to be transformed
-   * @returns {string}
-   * @private
-   */
-  static idToBinary(num) {
-    let bin = '';
-    let high = parseInt(num.slice(0, -10)) || 0;
-    let low = parseInt(num.slice(-10));
-    while (low > 0 || high > 0) {
-      bin = String(low & 1) + bin;
-      low = Math.floor(low / 2);
-      if (high > 0) {
-        low += 5_000_000_000 * (high % 2);
-        high = Math.floor(high / 2);
-      }
-    }
-    return bin;
-  }
-
-  /**
-   * Transforms a snowflake from a bit string to a decimal string.
-   * @param {string} num Bit string to be transformed
-   * @returns {Snowflake}
-   * @private
-   */
-  static binaryToId(num) {
-    let dec = '';
-
-    while (num.length > 50) {
-      const high = parseInt(num.slice(0, -32), 2);
-      const low = parseInt((high % 10).toString(2) + num.slice(-32), 2);
-
-      dec = (low % 10).toString() + dec;
-      num =
-        Math.floor(high / 10).toString(2) +
-        Math.floor(low / 10)
-          .toString(2)
-          .padStart(32, '0');
-    }
-
-    num = parseInt(num, 2);
-    while (num > 0) {
-      dec = (num % 10).toString() + dec;
-      num = Math.floor(num / 10);
-    }
-
-    return dec;
-  }
-
-  /**
    * Breaks user, role and everyone/here mentions by adding a zero width space after every @ character
    * @param {string} str The string to sanitize
    * @returns {string}