**Please describe the changes this PR makes and why it should be merged:**

Removes two unneeded private utilities (`Util.binaryToId` and `Util.idToBinary`) and increases the performance of `SnowflakeUtil.generate` and `SnowflakeUtil.deconstruct`:

| Iterations | Old `deconstruct` | New `deconstruct` |     Diff | Old `generate` | New `generate` |      Diff |
| :--------: | :---------------: | :---------------: | :------: | :------------: | :------------: | :-------: |
|    `10000` |           44.40ms |           26.15ms | **1.70** |        97.67ms |        14.64ms |  **6.67** |
|   `100000` |          400.57ms |          209.11ms | **1.92** |       746.16ms |        85.01ms |  **8.78** |
|  `1000000` |             3.68s |             2.07s | **1.78** |          7.72s |       479.30ms | **16.10** |

`SnowflakeUtil.deconstruct` is used in a total of 17 places: `ApplicationCommand`, `BaseGuild`, `Channel`, `Emoji`, `GuildAuditLogs`, `GuildPreview`, `Interaction`, `Message`, `Role`, `StageInstance`, `Sticker`, `StickerPack`, `Team`, `User`, `Webhook`, `Application`, and `TextBasedChannel`, to get the timestamp from a snowflake ID given by Discord. `Snowflake.generate` is used to generate the nonce in `GuildMemberManager`.

I could also make `SnowflakeUtil.getTimestamp(Snowflake)` to offer a +4 times faster alternative to `SnowflakeUtil.deconstruct(Snowflake).timestamp`, skipping allocations and a lot of overhead.

**Status and versioning classification:**

- Code changes have been tested against the Discord API, or there are no code changes
- I know how to update typings and have done so, or typings don't need updating