@wasdennnoch 2021-11-26
> `Snowflake.generate`, however, is not used by the library, but will grant large performance boosts for end-users who use it.

Sidenote, it actually is used here:
https://github.com/discordjs/discord.js/blob/54f937d82c4159d85e348dfb63b28f9f0c554805/src/managers/GuildMemberManager.js#L398